+++ app/controllers/accessibility_statement_controller.rb
@@ -1,3 +1,5 @@
 class AccessibilityStatementController < ApplicationController
+  include LocaleSwitchable
+  before_action :update_locale
   def index; end
 end
