+++ app/controllers/admin/admin_base_controller.rb
@@ -1,5 +1,6 @@
 module Admin
   class AdminBaseController < ApplicationController
+    include JourneyTypeIdentifiable
     before_action :check_vpn_ipaddr, :authenticate_admin_user!
     layout 'admin'.freeze
