+++ app/controllers/contacts_controller.rb
@@ -1,3 +1,5 @@
 class ContactsController < ApplicationController
+  include LocaleSwitchable
+  before_action :update_locale
   def show; end
 end
