+++ app/controllers/errors_controller.rb
@@ -1,4 +1,6 @@
 class ErrorsController < ApplicationController
+  include LocaleSwitchable
+  before_action :update_locale
   def show; end
 
   def page_not_found
