+++ app/controllers/privacy_policy_controller.rb
@@ -1,3 +1,5 @@
 class PrivacyPolicyController < ApplicationController
+  include LocaleSwitchable
+  before_action :update_locale
   def index; end
 end
