+++ app/controllers/problem_controller.rb
@@ -1,3 +1,5 @@
 class ProblemController < ApplicationController
+  include LocaleSwitchable
+  before_action :update_locale
   def index; end
 end
