+++ app/helpers/page_template_helper.rb
@@ -53,6 +53,7 @@ module PageTemplateHelper
     content_for(:navigation) { back_link(**back_link) unless back_link == :none }
     page_title_possibly_with_error({ page_title: page_title, head_title: head_title }, show_errors_for&.errors)
     content = capture(&content) if content
+    content_for(:language_switcher) { language_links if journey_type == :citizens }
     render(
       "shared/page_templates/#{template}_page_template",
       page_title: page_title,
